create class with method who return custom Predicate of Criteria api
